Extract transform iterator.
