+++ src/common/stats.h
@@ -8,7 +8,8 @@
 #include <limits>
 #include <vector>
 
-#include "common.h"  // AssertGPUSupport
+#include "common.h"              // AssertGPUSupport
+#include "transform_iterator.h"  // MakeIndexTransformIter
 #include "xgboost/generic_parameters.h"
 #include "xgboost/linalg.h"
