+++ src/data/ellpack_page.cu
@@ -7,6 +7,7 @@
 #include "../common/categorical.h"
 #include "../common/hist_util.cuh"
 #include "../common/random.h"
+#include "../common/transform_iterator.h"  // MakeIndexTransformIter
 #include "./ellpack_page.cuh"
 #include "device_adapter.cuh"
 #include "gradient_index.h"
